+++ src/ui/Windows/AddEdit_PropertySheet.cpp
@@ -500,12 +500,12 @@ void CAddEdit_PropertySheet::SetupInitialValues()
     m_AEMD.pcore->GetAllDependentEntries(original_uuid, dependentslist,
                                   m_AEMD.pci_original->IsAliasBase() ?
                                   CItemData::ET_ALIAS : CItemData::ET_SHORTCUT);
-    int num_dependents = dependentslist.size();
+    size_t num_dependents = dependentslist.size();
     if (num_dependents > 0) {
       m_AEMD.pcore->SortDependents(dependentslist, csDependents);
     }
 
-    m_AEMD.num_dependents = num_dependents;
+    m_AEMD.num_dependents = (int)num_dependents;
     m_AEMD.original_entrytype = entrytype;
     m_AEMD.dependents = CSecString(csDependents);
     dependentslist.clear();
